Refactor and DRY up handling request and response header matchers
