+++ Sources/Definitions/Request.swift
@@ -62,9 +62,9 @@ extension Request: Encodable {
 		self.query = query
 		self.headers = headers
 
-		let queryValues = Request.process(element: query, for: .query)
-		let headersValues = Request.process(element: headers, for: .header)
-		let bodyValues = Request.process(element: body, for: .body)
+		let queryValues = Toolbox.process(element: query, for: .query)
+		let headersValues = Toolbox.process(element: headers, for: .header)
+		let bodyValues = Toolbox.process(element: body, for: .body)
 
 		self.bodyEncoder = {
 			var container = $0.container(keyedBy: CodingKeys.self)
@@ -73,7 +73,7 @@ extension Request: Encodable {
 			if let query = queryValues?.node { try container.encode(query, forKey: .query) }
 			if let headers = headersValues?.node { try container.encode(headers, forKey: .headers) }
 			if let encodableBody = bodyValues?.node { try container.encode(encodableBody, forKey: .body) }
-			if let matchingRules = Request.mergeMatchingRulesFor(query: queryValues?.rules, header: headersValues?.rules, body: bodyValues?.rules) {
+			if let matchingRules = Toolbox.mergeMatchingRulesFor(body: bodyValues?.rules, query: queryValues?.rules, header: headersValues?.rules) {
 				try container.encode(matchingRules, forKey: .matchingRules)
 			}
 			if let generators = bodyValues?.generators { try container.encode(generators, forKey: .generators) }
@@ -82,41 +82,6 @@ extension Request: Encodable {
 
 	public func encode(to encoder: Encoder) throws {
 		try bodyEncoder(encoder)
-	}
-
-}
-
-private extension Request {
-
-	static func mergeMatchingRulesFor(query: AnyEncodable?, header: AnyEncodable?, body: AnyEncodable?) -> AnyEncodable? {
-		var merged: [String: AnyEncodable] = [:]
-
-		if let header = header {
-			merged["header"] = header
-		}
-
-		if let body = body {
-			merged["body"] = body
-		}
-
-		if let query = query {
-			merged["query"] = query
-		}
-
-		return merged.isEmpty ? nil : AnyEncodable(merged)
-	}
-
-	static func process(element: Any?, for interactionElement: PactInteractionNode) -> (node: AnyEncodable?, rules: AnyEncodable?, generators: AnyEncodable?)? {
-		if let element = element {
-			do {
-				let encodedElement = try PactBuilder(with: element).encoded(for: interactionElement)
-				return (node: encodedElement.node, rules: encodedElement.rules, generators: encodedElement.generators)
-			} catch {
-				fatalError("Can not process \(interactionElement.rawValue) with non-encodable (non-JSON safe) values")
-			}
-		}
-
-		return nil
 	}
 
 }
